+++ clipper-host/scripts/register.bat
@@ -17,6 +17,6 @@ exit /b
 
 :run
 
-reg add HKLM\SOFTWARE\Google\Chrome\NativeMessagingHosts\com.pushpin.pushpin /f /ve /t REG_SZ /d %~dp0win-com.pushpin.pushpin.json
+reg add HKLM\SOFTWARE\Google\Chrome\NativeMessagingHosts\com.pushpin.pushpin /f /ve /t REG_SZ /d %~dp0\..\dist\com.pushpin.pushpin.json
 
 pause
