+++ clipper-host/scripts/template.com.pushpin.pushpin.json
