+++ clipper-host/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "lib": [
+      "es2015",
+      "es2017",
+    ],
+    "outDir": "./dist/",
+  },
+  "files": [
+    "./src/index.ts"
+  ]
+}
