--- tsconfig.clipper-host.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "skipLibCheck": true,
-    "strict": true,
-    "noImplicitAny": false,
-    "sourceMap": true,
-    "noFallthroughCasesInSwitch": true,
-    "lib": [
-      "es2015",
-      "es2017",
-    ],
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "pretty": true,
-    "experimentalDecorators": true,
-    "outDir": "./clipper-host/dist/",
-    "importHelpers": true,
-  },
-  "files": [
-    "./clipper-host/src/host.ts"
-  ]
-}
